+++ src/triples.rs
@@ -46,7 +46,7 @@ impl TripleSect {
         }
     }
 
-    pub fn triples_with_id(&self, id: usize, id_kind: &IdKind) -> Box<dyn Iterator<Item = TripleId> + '_> {
+    pub fn triples_with_id(&self, id: u32, id_kind: &IdKind) -> Box<dyn Iterator<Item = TripleId> + '_> {
         match self {
             TripleSect::Bitmap(bitmap) => match id_kind {
                 IdKind::Subject => Box::new(BitmapIter::with_s(bitmap, id)),
@@ -108,10 +108,10 @@ impl OpIndex {
     pub fn size_in_bytes(&self) -> usize {
         self.sequence.len() * self.sequence.width() / 8 + self.bitmap.size_in_bytes()
     }
-    pub fn find(&self, o: usize) -> usize {
+    pub fn find(&self, o: u32) -> usize {
         self.bitmap.dict.select1(o as u64 - 1).unwrap() as usize
     }
-    pub fn last(&self, o: usize) -> usize {
+    pub fn last(&self, o: u32) -> usize {
         match self.bitmap.dict.select1(o as u64) {
             Some(index) => index as usize - 1,
             None => self.bitmap.dict.len() - 1,
@@ -142,11 +142,11 @@ impl TriplesBitmap {
         self.adjlist_z.size_in_bytes() + self.op_index.size_in_bytes() + self.wavelet_y.size_in_bytes()
     }
 
-    pub fn find_y(&self, subject_id: usize) -> usize {
+    pub fn find_y(&self, subject_id: u32) -> u32 {
         if subject_id == 0 {
             return 0;
         }
-        self.bitmap_y.dict.select1(subject_id as u64 - 1).unwrap() as usize + 1
+        self.bitmap_y.dict.select1(subject_id as u64 - 1).unwrap() as u32 + 1
     }
 
     pub fn read<R: BufRead>(reader: &mut R, triples_ci: &ControlInfo) -> io::Result<Self> {
@@ -217,7 +217,7 @@ impl TriplesBitmap {
         Ok(TriplesBitmap { order, bitmap_y, adjlist_z, op_index, wavelet_y })
     }
 
-    pub fn coord_to_triple(&self, x: usize, y: usize, z: usize) -> io::Result<TripleId> {
+    pub fn coord_to_triple(&self, x: u32, y: u32, z: u32) -> io::Result<TripleId> {
         use io::Error;
         use io::ErrorKind::InvalidData;
         if x == 0 || y == 0 || z == 0 {
@@ -251,12 +251,12 @@ pub struct BitmapIter<'a> {
     // triples data
     triples: &'a TriplesBitmap,
     // x-coordinate identifier
-    x: usize,
+    x: u32,
     // current position
-    pos_y: usize,
-    pos_z: usize,
-    max_y: usize,
-    max_z: usize,
+    pos_y: u32,
+    pos_z: u32,
+    max_y: u32,
+    max_z: u32,
 }
 
 impl<'a> BitmapIter<'a> {
@@ -272,7 +272,7 @@ impl<'a> BitmapIter<'a> {
     }
 
     /// see <https://github.com/rdfhdt/hdt-cpp/blob/develop/libhdt/src/triples/BitmapTriplesIterators.cpp>
-    pub fn with_s(triples: &'a TriplesBitmap, subject_id: usize) -> Self {
+    pub fn with_s(triples: &'a TriplesBitmap, subject_id: u32) -> Self {
         let min_y = triples.find_y(subject_id - 1);
         let min_z = triples.adjlist_z.find(min_y);
         let max_y = triples.find_y(subject_id);
@@ -293,7 +293,7 @@ impl<'a> Iterator for BitmapIter<'a> {
             return None;
         }
 
-        let y = self.triples.wavelet_y.get(self.pos_y);
+        let y = self.triples.wavelet_y.get(self.pos_y as usize);
         let z = self.triples.adjlist_z.get_id(self.pos_z);
         let triple_id = self.triples.coord_to_triple(self.x, y, z).unwrap();
 
@@ -311,15 +311,17 @@ impl<'a> Iterator for BitmapIter<'a> {
     }
 }
 
+/// Type for a triple encoded as numeric IDs for subject, predicate and object, respectively.
+/// See <https://www.rdfhdt.org/hdt-binary-format/#triples>.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
 pub struct TripleId {
-    pub subject_id: usize,
-    pub predicate_id: usize,
-    pub object_id: usize,
+    pub subject_id: u32,
+    pub predicate_id: u32,
+    pub object_id: u32,
 }
 
 impl TripleId {
-    pub const fn new(subject_id: usize, predicate_id: usize, object_id: usize) -> Self {
+    pub const fn new(subject_id: u32, predicate_id: u32, object_id: u32) -> Self {
         TripleId { subject_id, predicate_id, object_id }
     }
 }
